更新 数据库实体

+++ src/main/java/com/njust/csa/reg/repository/docker/TableInfoRepo.java
@@ -3,6 +3,6 @@ package com.njust.csa.reg.repository.docker;
 import com.njust.csa.reg.repository.entities.TableInfoEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TableInfoRepo extends CrudRepository<TableInfoEntity, Integer> {
+public interface TableInfoRepo extends CrudRepository<TableInfoEntity, Long> {
 
 }
